Files were moved and other changes
