--- love/src/components/MainTel/Scheduler/Scheduler.module.css
@@ -1,43 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.allComponentes {
-  /* display: inline-grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    grid-row-gap: 0.5em;
-    grid-column-gap: 1em;
-    align-items: center;
-    padding: 1em 2em;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 0.5em; */
-  width: 100%;
-  /* margin: 0 auto; */
-  display: flex;
-}
-
-.leftDiv {
-  float: left;
-  width: 25%;
-  padding: 0.5em;
-  height: 95%;
-}
-
-.middleDiv {
-  float: left;
-  width: 50%;
-  padding: 0.5em;
-  height: 95%;
-}
-
-.rigthDiv {
-  float: right;
-  width: 25%;
-  padding: 0.5em;
-  height: 95%;
-}
