+++ love/src/redux/selectors/selectors.js
@@ -1194,7 +1194,7 @@ export const getDetailedState = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-detailedState`];
   const summaryData = getStreamsData(state, subscriptions);
   return {
-    subState: summaryData[`event-Scheduler-${salindex}-detailedState`]?.[0]?.subState.value ?? 0,
+    subState: summaryData[`event-Scheduler-${salindex}-detailedState`]?.[0]?.subState?.value ?? 0,
   }
 };
 
@@ -1202,8 +1202,8 @@ export const getObservingMode = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-observingMode`];
   const observingMode = getStreamsData(state, subscriptions);
   return {
-    mode: observingMode[`event-Scheduler-${salindex}-observingMode`]?.[0]?.mode.value ?? '',
-    type: observingMode[`event-Scheduler-${salindex}-observingMode`]?.[0]?.type.value ?? '',
+    mode: observingMode[`event-Scheduler-${salindex}-observingMode`]?.[0]?.mode?.value ?? '',
+    type: observingMode[`event-Scheduler-${salindex}-observingMode`]?.[0]?.type?.value ?? '',
   }
 };
 
@@ -1211,10 +1211,10 @@ export const getGeneralInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-generalInfo`];
   const generalInfo = getStreamsData(state, subscriptions);
   return {
-    isNigth: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.isNigth.value ?? false, 
-    night: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.nigth.value ?? 0,
-    sunset: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.sunset.value ?? 1679698547,
-    sunrise: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.sunrise.value ?? 1679738147,
+    isNigth: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.isNigth?.value ?? false, 
+    night: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.nigth?.value ?? 0,
+    sunset: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.sunset?.value ?? 1679698547,
+    sunrise: generalInfo[`event-Scheduler-${salindex}-generalInfo`]?.[0]?.sunrise?.value ?? 1679738147,
   }
 };
 
@@ -1222,9 +1222,9 @@ export const getFilterSwap = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-needFilterSwap`];
   const filterSwap = getStreamsData(state, subscriptions);
   return {
-    needSwap: filterSwap[`event-Scheduler-${salindex}-needFilterSwap`]?.[0]?.needSwap.value ?? false,
-    filterToMount: filterSwap[`event-Scheduler-${salindex}-needFilterSwap`]?.[0]?.filterToMount.value ?? '',
-    filterToUnmount: filterSwap[`event-Scheduler-${salindex}-needFilterSwap`]?.[0]?.filterToUnmount.value ?? '',
+    needSwap: filterSwap[`event-Scheduler-${salindex}-needFilterSwap`]?.[0]?.needSwap?.value ?? false,
+    filterToMount: filterSwap[`event-Scheduler-${salindex}-needFilterSwap`]?.[0]?.filterToMount?.value ?? '',
+    filterToUnmount: filterSwap[`event-Scheduler-${salindex}-needFilterSwap`]?.[0]?.filterToUnmount?.value ?? '',
   }
 };
 
@@ -1232,27 +1232,27 @@ export const getObservatoryStatus = (state, salindex) => {
   const subscriptions = [`telemetry-Scheduler-${salindex}-observatoryState`, `event-Scheduler-${salindex}-target`];
   const observatoryStatus = getStreamsData(state, subscriptions);
   return {
-    pointingRa: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.ra.value ?? 0.0,
-    pointingDecl: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.declination.value ?? 0.0,
-    pointingPosAngle: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.positionAngle.value ?? 0.0,
-    pointingParallAngle: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.parallacticAngle.value ?? 0.0,
-    simonyiTracking: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.tacking.value ?? false,
-    simonyiAl: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.telescopeAltitude.value ?? 0.0,
-    simonyiAz: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.telescopeAzimuth.value ?? 0.0,
-    simonyiRot: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.telescopeRotator.value ?? 0.0,
-    domeAlt: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.domeAltitude.value ?? 0.0,
-    domeAz: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.[0]?.domeAzimuth.value ?? 0,
-    moonRa: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonRa.value ?? 0.0,
-    moonDec: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonDec.value ?? 0.0,
-    moonAlt: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonAlt.value ?? 0.0,
-    moonAz: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonAz.value ?? 0.0,
-    moonDistance: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonDistance.value ?? 0.0,
-    moonPhase: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonPhase.value ?? 0.0,
-    sunRa: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunRa.value ?? 0.0,
-    sunDec: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunDec.value ?? 0.0,
-    sunAlt: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunAlt.value ?? 0.0,
-    sunAz: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunAz.value ?? 0.0,
-    solarElong: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.solarElong.value ?? 0.0,
+    pointingRa: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.ra?.value ?? 0.0,
+    pointingDecl: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.declination?.value ?? 0.0,
+    pointingPosAngle: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.positionAngle?.value ?? 0.0,
+    pointingParallAngle: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.parallacticAngle?.value ?? 0.0,
+    simonyiTracking: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.tacking?.value ?? false,
+    simonyiAl: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.telescopeAltitude?.value ?? 0.0,
+    simonyiAz: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.telescopeAzimuth?.value ?? 0.0,
+    simonyiRot: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.telescopeRotator?.value ?? 0.0,
+    domeAlt: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.domeAltitude?.value ?? 0.0,
+    domeAz: observatoryStatus[`telemetry-Scheduler-${salindex}-observatoryState`]?.domeAzimuth?.value ?? 0,
+    moonRa: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonRa?.value ?? 0.0,
+    moonDec: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonDec?.value ?? 0.0,
+    moonAlt: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonAlt?.value ?? 0.0,
+    moonAz: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonAz?.value ?? 0.0,
+    moonDistance: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonDistance?.value ?? 0.0,
+    moonPhase: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.moonPhase?.value ?? 0.0,
+    sunRa: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunRa?.value ?? 0.0,
+    sunDec: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunDec?.value ?? 0.0,
+    sunAlt: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunAlt?.value ?? 0.0,
+    sunAz: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.sunAz?.value ?? 0.0,
+    solarElong: observatoryStatus[`event-Scheduler-${salindex}-target`]?.[0]?.solarElong?.value ?? 0.0,
   }
 };
 
@@ -1260,23 +1260,23 @@ export const getCurrentTargetInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-target`];
   const currentTarget =  getStreamsData(state, subscriptions);
   return {
-    currentTargetId: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.targetId.value ?? 0,
-    currentRequestTime: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.requestTime.value ?? 0,
-    currentRequestMjd: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.requestMjd.value ?? 0,
-    currentRa: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.ra.value ?? 0,
-    currentDecl: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.decl.value ?? 0,
-    currentSkyAngle: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.skyAngle.value ?? 0,
-    currentFilter: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.filter.value ?? '',
-    currentNumExposures: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.numExposures.value ?? 0,
-    currentExposureTimes: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.exposureTimes.value ?? 0,
-    currentSlewTime: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.slewTime.value ?? 0,
-    currentOffsetX: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.offsetX.value ?? 0,
-    currentOffsetY: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.offsetY.value ?? 0,
-    currentNumProposals: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.numProposals.value ?? 0,
-    currentProposalId: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.proposalId.value ?? [],
-    currentSequenceDuration: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.sequenceDuration.value ?? 0,
-    currentSequenceNVisits: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.sequenceNVisits.value ?? 0,
-    currentSequenceVisits: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.sequenceVisits.value ?? 0,
+    currentTargetId: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.targetId?.value ?? 0,
+    currentRequestTime: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.requestTime?.value ?? 0,
+    currentRequestMjd: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.requestMjd?.value ?? 0,
+    currentRa: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.ra?.value ?? 0,
+    currentDecl: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.decl?.value ?? 0,
+    currentSkyAngle: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.skyAngle?.value ?? 0,
+    currentFilter: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.filter?.value ?? '',
+    currentNumExposures: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.numExposures?.value ?? 0,
+    currentExposureTimes: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.exposureTimes?.value ?? 0,
+    currentSlewTime: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.slewTime?.value ?? 0,
+    currentOffsetX: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.offsetX?.value ?? 0,
+    currentOffsetY: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.offsetY?.value ?? 0,
+    currentNumProposals: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.numProposals?.value ?? 0,
+    currentProposalId: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.proposalId?.value ?? [],
+    currentSequenceDuration: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.sequenceDuration?.value ?? 0,
+    currentSequenceNVisits: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.sequenceNVisits?.value ?? 0,
+    currentSequenceVisits: currentTarget[`event-Scheduler-${salindex}-target`]?.[0]?.sequenceVisits?.value ?? 0,
   }
 };
 
@@ -1287,10 +1287,10 @@ export const getSkyMapInfo = (state, salindex) => {
   ];
   const skyMapInfo = getStreamsData(state, subscriptions);
   return {
-    rotSkyPos: skyMapInfo[`event-Scheduler-${salindex}-observation`]?.[0]?.rotSkyPos.value ?? 0,
-    predictedTargetsRa: skyMapInfo[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.ra.value ?? [],
-    predictedTargetsDecl: skyMapInfo[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.decl.value ?? [],
-    predictedTargetsRotSkyPos: skyMapInfo[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.rotSkyPos.value ?? [],
+    rotSkyPos: skyMapInfo[`event-Scheduler-${salindex}-observation`]?.[0]?.rotSkyPos?.value ?? 0,
+    predictedTargetsRa: skyMapInfo[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.ra?.value ?? [],
+    predictedTargetsDecl: skyMapInfo[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.decl?.value ?? [],
+    predictedTargetsRotSkyPos: skyMapInfo[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.rotSkyPos?.value ?? [],
   }
 };
 
@@ -1298,15 +1298,15 @@ export const lastTargetInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-observation`];
   const lastTarget = getStreamsData(state, subscriptions);
   return {
-    lastTargetId: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.targetId.value ?? 0,
-    lastTargetRa: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.ra.value ?? 0,
-    lastTargetDecl: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.decl.value ?? 0,
-    lastTargetRotSkyPos: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.rotSkyPos.value ?? 0,
-    lastTargetMjd: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.mjd.value ?? 0,
-    lastTargetExpTime: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.exptime.value ?? 0,
-    lastTargetFilter: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.filter.value ?? '',
-    lastTargetNexp: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.nexp.value ?? 0,
-    lastTargetMoreInfo: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.additionalInformation.value ?? '',
+    lastTargetId: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.targetId?.value ?? 0,
+    lastTargetRa: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.ra?.value ?? 0,
+    lastTargetDecl: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.decl?.value ?? 0,
+    lastTargetRotSkyPos: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.rotSkyPos?.value ?? 0,
+    lastTargetMjd: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.mjd?.value ?? 0,
+    lastTargetExpTime: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.exptime?.value ?? 0,
+    lastTargetFilter: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.filter?.value ?? '',
+    lastTargetNexp: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.nexp?.value ?? 0,
+    lastTargetMoreInfo: lastTarget[`event-Scheduler-${salindex}-observation`]?.[0]?.additionalInformation?.value ?? '',
   }
 };
 
@@ -1314,11 +1314,11 @@ export const nextTargetInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-timeToNextTarget`];
   const nextTarget = getStreamsData(state, subscriptions);
   return {
-    nextTargetCurrentTime: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.currentTime.value ?? 0,
-    nextTimeWaitTime: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.waitTime.value ?? 0,
-    nextTargetRa: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.ra.value ?? 0,
-    nextTargetDecl: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.decl.value ?? 0,
-    nextTargetRotSkyPos: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.rotSkyPos.value ?? 0,
+    nextTargetCurrentTime: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.currentTime?.value ?? 0,
+    nextTimeWaitTime: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.waitTime?.value ?? 0,
+    nextTargetRa: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.ra?.value ?? 0,
+    nextTargetDecl: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.decl?.value ?? 0,
+    nextTargetRotSkyPos: nextTarget[`event-Scheduler-${salindex}-timeToNextTarget`]?.[0]?.rotSkyPos?.value ?? 0,
   }
 };
 
@@ -1326,14 +1326,14 @@ export const predictedTargetsInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-predictedSchedule`];
   const predictedTargets = getStreamsData(state, subscriptions);
   return {
-    predTargetsNumTargets: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.numberOfTargets.value ?? 0,
-    predTargetsRa: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.ra.value ?? [],
-    predTargetsDecl: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.decl.value ?? [],
-    predTargetsRotSkyPos: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.rotSkyPos.value ?? [],
-    predTargetsMjd: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.mjd.value ?? [],
-    predTargetsExpTime: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.exptime.value ?? [],
-    predTargetsInstrConfig: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.instrumentConfiguration.value ?? '',
-    predTargetsNexp: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.nexp.value ?? [],
+    predTargetsNumTargets: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.numberOfTargets?.value ?? 0,
+    predTargetsRa: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.ra?.value ?? [],
+    predTargetsDecl: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.decl?.value ?? [],
+    predTargetsRotSkyPos: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.rotSkyPos?.value ?? [],
+    predTargetsMjd: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.mjd?.value ?? [],
+    predTargetsExpTime: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.exptime?.value ?? [],
+    predTargetsInstrConfig: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.instrumentConfiguration?.value ?? '',
+    predTargetsNexp: predictedTargets[`event-Scheduler-${salindex}-predictedSchedule`]?.[0]?.nexp?.value ?? [],
   }
 };
 
@@ -1341,10 +1341,10 @@ export const getSurveysInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-surveyTopology`];
   const surveys = getStreamsData(state, subscriptions);
   return {
-    surveysNumGenProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.numGeneralProps.value ?? 0,
-    surveysGenProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.generalPropos.value ?? '',
-    surveysNumSeqProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.numSeqProps.value ?? 0,
-    surveysSeqProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.generalPropos.value ?? '',
+    surveysNumGenProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.numGeneralProps?.value ?? 0,
+    surveysGenProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.generalPropos?.value ?? '',
+    surveysNumSeqProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.numSeqProps?.value ?? 0,
+    surveysSeqProps: surveys[`event-Scheduler-${salindex}-surveyTopology`]?.[0]?.generalPropos?.value ?? '',
   }
 };
 
@@ -1352,15 +1352,15 @@ export const getBlocksInfo = (state, salindex) => {
   const subscriptions = [`event-Scheduler-${salindex}-blockInventory`, `event-Scheduler-${salindex}-blockStatus`];
   const blocks = getStreamsData(state, subscriptions);
   return {
-    blockInvId: blocks[`event-Scheduler-${salindex}-blockInventory`]?.[0]?.ids.vallue ?? '',
-    blockInvStatus: blocks[`event-Scheduler-${salindex}-blockInventory`]?.[0]?.status.value ?? '',
-    blockId: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.id.value ?? '',
-    blockStatusId: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.statusId.value ?? 0,
-    blockStatus: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.status.value ?? '',
-    blockExecCompl: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.executionsCompleted.value ?? 0,
-    blockExecTotal: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.executionsTotal.value ?? 0,
-    blockHash: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.hash.value ?? '',
-    blockDef: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.definition.value ?? '',
+    blockInvId: blocks[`event-Scheduler-${salindex}-blockInventory`]?.[0]?.ids?.vallue ?? '',
+    blockInvStatus: blocks[`event-Scheduler-${salindex}-blockInventory`]?.[0]?.status?.value ?? '',
+    blockId: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.id?.value ?? '',
+    blockStatusId: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.statusId?.value ?? 0,
+    blockStatus: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.status?.value ?? '',
+    blockExecCompl: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.executionsCompleted?.value ?? 0,
+    blockExecTotal: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.executionsTotal?.value ?? 0,
+    blockHash: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.hash?.value ?? '',
+    blockDef: blocks[`event-Scheduler-${salindex}-blockStatus`]?.[0]?.definition?.value ?? '',
   }
 };
